PSR-4 : extract fixtures
