+++ tests/Feature/Fixtures/EternalSupervisor.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Laravel\Horizon\Tests\Feature\Fixtures;
+
+class EternalSupervisor
+{
+    public $name = 'eternal';
+
+    public function terminate()
+    {
+        //
+    }
+
+    public function isRunning()
+    {
+        return true;
+    }
+}
