+++ tests/Feature/Fixtures/SupervisorProcessWithFakeRestart.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Laravel\Horizon\Tests\Feature\Fixtures;
+
+use Laravel\Horizon\SupervisorProcess;
+
+class SupervisorProcessWithFakeRestart extends SupervisorProcess
+{
+    public $wasRestarted = false;
+
+    public function restart()
+    {
+        $this->wasRestarted = true;
+    }
+}
